update todo theme
